Document test sources.
